More adaptations.
--HG--
branch : 0.7
